rename notifications model
